requirements.txt
numpy==1.20.3
PyYAML==5.4.1
sumolib==1.9.2
traci==1.9.2
